常规修改：   (1): 完成删除提示。 bug 修改：    解决项目管理下时间不正确的问题

+++ src/app/business/department/organization-management/organization-management.component.ts
@@ -173,17 +173,18 @@ export class OrganizationManagementComponent implements OnInit {
       this.mustone = false;
       this.gtone = true;
     } else {
-      this.mustone = false;
-      this.openstatus = false;
-      for (let j = 0; j < haschecklen; j++) {
-        this.req.deleteOrganization('id=' +  this.datas[this.hasChecked[j]].id)
-          .subscribe(res => {
-            if (j === haschecklen - 1) {
-              this.resMessage = res.message;
-              this.status = Number(res.status);
-              this.Update();
-            }
-          });
+      if (this.commonfun.deleteChecked(this.datas, this.hasChecked, 'name')) {
+        this.openstatus = false;
+        for (let j = 0; j < haschecklen; j++) {
+          this.req.deleteOrganization('id=' +  this.datas[this.hasChecked[j]].id)
+            .subscribe(res => {
+              if (j === haschecklen - 1) {
+                this.resMessage = res.message;
+                this.status = Number(res.status);
+                this.Update();
+              }
+            });
+        }
       }
     }
   }
