+++ src/app/business/technology/technicspack-temperature/technicspack-temperature.component.ts
@@ -12,7 +12,7 @@ import {CommonfunService} from '../../../shared/commonfun.service';
 })
 export class TechnicspackTemperatureComponent implements OnInit {
   public technologyParamsPackWordList: Array<TechnologyParamsPackWord>;
-  public datas: Array<TechnologyTemperatureQueryList>
+  public datas: Array<TechnologyTemperatureQueryList>;
   public modalRef: BsModalRef;
   public pageBody: PageBody;
   public num: number;
@@ -28,13 +28,13 @@ export class TechnicspackTemperatureComponent implements OnInit {
   public gtone: boolean;
   public resMessage: string;
   public listenDescModal: boolean;
-  constructor(
-    private modalService: BsModalService,
-    private req: ReqService,
-    private fb: FormBuilder,
-    private commonfun: CommonfunService
-  ) {
+
+  constructor(private modalService: BsModalService,
+              private req: ReqService,
+              private fb: FormBuilder,
+              private commonfun: CommonfunService) {
   }
+
   ngOnInit() {
     this.status = 0;
     this.openstatus = true;
@@ -95,31 +95,32 @@ export class TechnicspackTemperatureComponent implements OnInit {
     });
     this.technologyParamsPackWordList = [
       new TechnologyParamsPackWord('方 案 名 称', 'name', '', ''),
-      new TechnologyParamsPackWord('铝板厚度',	'al_thickness',	'毫米', '将（在）生产铝板厚度'),
-      new TechnologyParamsPackWord('铝板宽度',	'al_width',	'毫米'	, '将（在）生产铝板宽度'),
+      new TechnologyParamsPackWord('铝板厚度', 'al_thickness', '毫米', '将（在）生产铝板厚度'),
+      new TechnologyParamsPackWord('铝板宽度', 'al_width', '毫米', '将（在）生产铝板宽度'),
       new TechnologyParamsPackWord('一涂一区温度', 'temperature_1_1', '	摄氏度', '一涂一区温度设定'),
-      new TechnologyParamsPackWord('一涂一区温度差值',	'temperature_1_1_d', '	摄氏度', '一涂一区温度安全值设定'),
-      new TechnologyParamsPackWord('一涂二区温度',	'temperature_1_2', '	摄氏度', '一涂二区温度设定'),
-      new TechnologyParamsPackWord('一涂二区温度差值',	'temperature_1_2_d', '	摄氏度', '一涂二区温度安全值设定'),
-      new TechnologyParamsPackWord('一涂三区温度',	'temperature_1_3', '	摄氏度', '一涂三区温度设定'),
-      new TechnologyParamsPackWord('一涂三区温度差值',	'temperature_1_3_d', '	摄氏度', '一涂三区温度安全值设定'),
-      new TechnologyParamsPackWord('一涂四区温度',	'temperature_1_4', '	摄氏度', '一涂四区温度设定'),
-      new TechnologyParamsPackWord('一涂四区温度差值',	'temperature_1_4_d', '	摄氏度', '一涂四区温度安全值设定'),
-      new TechnologyParamsPackWord('一涂五区温度',	'temperature_1_5', '	摄氏度', '一涂五区温度设定'),
-      new TechnologyParamsPackWord('一涂五区温度差值',	'temperature_1_5_d', '	摄氏度', '一涂五区温度安全值设定'),
-      new TechnologyParamsPackWord('二涂一区温度',	'temperature_2_1', '	摄氏度', '二涂一区温度设定'),
-      new TechnologyParamsPackWord('二涂一区温度差值',	'temperature_2_1_d', '	摄氏度', '二涂一区温度安全值设定'),
-      new TechnologyParamsPackWord('二涂二区温度',	'temperature_2_2', '	摄氏度', '二涂二区温度设定'),
-      new TechnologyParamsPackWord('二涂二区温度差值',	'temperature_2_2_d', '	摄氏度', '二涂二区温度安全值设定'),
-      new TechnologyParamsPackWord('二涂三区温度',	'temperature_2_3', '	摄氏度', '二涂三区温度设定'),
-      new TechnologyParamsPackWord('二涂三区温度差值',	'temperature_2_3_d', '	摄氏度', '二涂三区温度安全值设定'),
-      new TechnologyParamsPackWord('二涂四区温度',	'temperature_2_4', '	摄氏度', '二涂四区温度设定'),
-      new TechnologyParamsPackWord('二涂四区温度差值',	'temperature_2_4_d', '	摄氏度', '二涂四区温度安全值设定'),
-      new TechnologyParamsPackWord('二涂五区温度',	'temperature_2_5', '	摄氏度', '二涂五区温度设定'),
-      new TechnologyParamsPackWord('二涂五区温度差值',	'temperature_2_5_d', '	摄氏度', '二涂五区温度安全值设定'),
+      new TechnologyParamsPackWord('一涂一区温度差值', 'temperature_1_1_d', '	摄氏度', '一涂一区温度安全值设定'),
+      new TechnologyParamsPackWord('一涂二区温度', 'temperature_1_2', '	摄氏度', '一涂二区温度设定'),
+      new TechnologyParamsPackWord('一涂二区温度差值', 'temperature_1_2_d', '	摄氏度', '一涂二区温度安全值设定'),
+      new TechnologyParamsPackWord('一涂三区温度', 'temperature_1_3', '	摄氏度', '一涂三区温度设定'),
+      new TechnologyParamsPackWord('一涂三区温度差值', 'temperature_1_3_d', '	摄氏度', '一涂三区温度安全值设定'),
+      new TechnologyParamsPackWord('一涂四区温度', 'temperature_1_4', '	摄氏度', '一涂四区温度设定'),
+      new TechnologyParamsPackWord('一涂四区温度差值', 'temperature_1_4_d', '	摄氏度', '一涂四区温度安全值设定'),
+      new TechnologyParamsPackWord('一涂五区温度', 'temperature_1_5', '	摄氏度', '一涂五区温度设定'),
+      new TechnologyParamsPackWord('一涂五区温度差值', 'temperature_1_5_d', '	摄氏度', '一涂五区温度安全值设定'),
+      new TechnologyParamsPackWord('二涂一区温度', 'temperature_2_1', '	摄氏度', '二涂一区温度设定'),
+      new TechnologyParamsPackWord('二涂一区温度差值', 'temperature_2_1_d', '	摄氏度', '二涂一区温度安全值设定'),
+      new TechnologyParamsPackWord('二涂二区温度', 'temperature_2_2', '	摄氏度', '二涂二区温度设定'),
+      new TechnologyParamsPackWord('二涂二区温度差值', 'temperature_2_2_d', '	摄氏度', '二涂二区温度安全值设定'),
+      new TechnologyParamsPackWord('二涂三区温度', 'temperature_2_3', '	摄氏度', '二涂三区温度设定'),
+      new TechnologyParamsPackWord('二涂三区温度差值', 'temperature_2_3_d', '	摄氏度', '二涂三区温度安全值设定'),
+      new TechnologyParamsPackWord('二涂四区温度', 'temperature_2_4', '	摄氏度', '二涂四区温度设定'),
+      new TechnologyParamsPackWord('二涂四区温度差值', 'temperature_2_4_d', '	摄氏度', '二涂四区温度安全值设定'),
+      new TechnologyParamsPackWord('二涂五区温度', 'temperature_2_5', '	摄氏度', '二涂五区温度设定'),
+      new TechnologyParamsPackWord('二涂五区温度差值', 'temperature_2_5_d', '	摄氏度', '二涂五区温度安全值设定'),
     ];
     this.Update();
   }
+
   // 控制模态框, 增，修，查
   public openModal(template: TemplateRef<any>, i): void {
     this.inputvalid = false;
@@ -146,7 +147,7 @@ export class TechnicspackTemperatureComponent implements OnInit {
           this.paramModifyForm.reset(this.detail.temperaturedata);
           this.modalRef = this.modalService.show(template);
           this.listenDescModal = false;
-        }else {
+        } else {
           this.mustone = true;
         }
       } else {
@@ -169,11 +170,13 @@ export class TechnicspackTemperatureComponent implements OnInit {
       this.modalRef = this.modalService.show(template);
     }
   }
+
   // 监控翻页事件
   public getPageBody(event): void {
     this.pageBody = event;
     this.Update();
   }
+
   // 全选 或 全不选
   public getAllCheckBoxStatus(e): void {
     if (e.srcElement.checked === true) {
@@ -185,13 +188,14 @@ export class TechnicspackTemperatureComponent implements OnInit {
       this.checked = '';
     }
   }
+
   // 得到已选择的checkBox
   public getCheckBoxStatus(e, i): void {
     let haschecklen = this.hasChecked.length;
     if (e.srcElement.checked === true) {
       this.hasChecked.push(i);
     } else {
-      for (let j = 0; j < haschecklen; j++ ) {
+      for (let j = 0; j < haschecklen; j++) {
         if (this.hasChecked[j] === i) {
           this.hasChecked.splice(j, 1);
         }
@@ -203,6 +207,7 @@ export class TechnicspackTemperatureComponent implements OnInit {
       this.detail = null;
     }
   }
+
   // 删除表格 并且 重新请求数据(不管删除多少条，只请求数据刷新一次)
   public delete(): void {
     let haschecklen = this.hasChecked.length;
@@ -210,23 +215,23 @@ export class TechnicspackTemperatureComponent implements OnInit {
       this.mustone = false;
       this.gtone = true;
     } else {
-      this.mustone = false;
-      this.openstatus = false;
-      for (let j = 0; j < haschecklen; j++) {
-        const body = 'al_thickness=' +  this.datas[this.hasChecked[j]]['althickness'] + '&&al_width=' + this.datas[this.hasChecked[j]]['alwidth'];
-        console.log(body);
-        this.req.DeleteTechnicsPackTemperature(body)
-          .subscribe(res => {
-            console.log(res);
-            if (j === haschecklen - 1) {
-              this.resMessage = res['message'];
-              this.status = Number(res.status);
-              this.Update();
-            }
-          });
+      if (this.commonfun.deleteChecked(this.datas, this.hasChecked, 'name')) {
+        this.openstatus = false;
+        for (let j = 0; j < haschecklen; j++) {
+          const body = 'al_thickness=' + this.datas[this.hasChecked[j]]['althickness'] + '&&al_width=' + this.datas[this.hasChecked[j]]['alwidth'];
+          this.req.DeleteTechnicsPackTemperature(body)
+            .subscribe(res => {
+              if (j === haschecklen - 1) {
+                this.resMessage = res['message'];
+                this.status = Number(res.status);
+                this.Update();
+              }
+            });
+        }
       }
     }
   }
+
   // 生产线的添加 并且 重新请求数据，防止增加的是第十一条表格
   public paramsAdd(): void {
     if (this.paramAddForm.valid) {
@@ -243,6 +248,7 @@ export class TechnicspackTemperatureComponent implements OnInit {
       this.inputvalid = true;
     }
   }
+
 //  修改表格内容
   public paramsModify(): void {
     if (this.paramModifyForm.valid) {
@@ -259,6 +265,7 @@ export class TechnicspackTemperatureComponent implements OnInit {
       this.inputvalid = true;
     }
   }
+
   // 刷新
   public Update(): void {
     this.gtone = false;
